refactor(tauri-shell): modularize shell runtime
