+++ apps/tauri-shell/src-tauri/src/backend/mod.rs
@@ -1,0 +1,6 @@
+mod health;
+mod paths;
+mod process;
+
+pub(crate) use self::health::wait_for_backend_ready;
+pub(crate) use self::process::{spawn_packaged_backend, BackendProcessState};
